Rewrite summary MVC classes
