+++ apps/desktop/desktop/src/main/java/bisq/desktop/main/content/user/accounts/create/data/payment_form/PaymentFormModel.java
@@ -22,4 +22,9 @@ import lombok.Getter;
 
 @Getter
 public abstract class PaymentFormModel implements Model {
+    protected final String id;
+
+    public PaymentFormModel(String id) {
+        this.id = id;
+    }
 }
