+++ i18n/src/main/resources/application.properties
@@ -348,7 +348,15 @@ user.paymentAccounts.sepa.bic=BIC
 user.paymentAccounts.sepa.acceptedCountries=Countries you accept payments from
 user.paymentAccounts.sepa.acceptAllSepaCountries=Accept payments from all SEPA countries
 user.paymentAccounts.sepa.acceptedCountriesHint=Select specific countries to limit payment sources"
-user.paymentAccounts.summary.overview=Overview
+
+
+user.paymentAccounts.summary.review=Review payment account
+user.paymentAccounts.summary.paymentMethod=Payment method
+user.paymentAccounts.summary.accountName=Account name
+user.paymentAccounts.summary.risk=Chargeback risk
+user.paymentAccounts.summary.tradeLimit=Trade Limit
+user.paymentAccounts.summary.accountDetails=Account details
+
 
 user.paymentAccounts.summary.country=Country
 user.paymentAccounts.summary.city=City
@@ -360,10 +368,10 @@ user.paymentAccounts.summary.iban=IBAN
 user.paymentAccounts.summary.bic=BIC
 user.paymentAccounts.summary.acceptedCountries=Accepted Countries
 
-user.paymentAccounts.summary.risk=Chargeback risk
+
 user.paymentAccounts.summary.risk.high=High
 user.paymentAccounts.summary.risk.medium=Medium
 user.paymentAccounts.summary.risk.low=Low
 user.paymentAccounts.summary.risk.unknown=Unknown
-user.paymentAccounts.summary.tradeLimit=Trade Limit
-user.paymentAccounts.summary.accountName=Account name
+
+
